+++ pkg/innerring/processors/netmap/processor.go
@@ -126,6 +126,11 @@ func New(p *Params) (*Processor, error) {
 		return nil, fmt.Errorf("ir/netmap: can't create worker pool: %w", err)
 	}
 
+	// Override any setting for v2, it's managed by contract.
+	if p.NetmapClient.IsNodeV2() {
+		p.CleanupEnabled = false
+	}
+
 	return &Processor{
 		log:            p.Log,
 		pool:           pool,
@@ -184,7 +189,7 @@ func (np *Processor) ListenerNotaryParsers() []event.NotaryParserInfo {
 	var (
 		p event.NotaryParserInfo
 
-		pp = make([]event.NotaryParserInfo, 0, 2)
+		pp = make([]event.NotaryParserInfo, 0, 3)
 	)
 
 	p.SetScriptHash(np.netmapClient.ContractAddress())
@@ -192,6 +197,11 @@ func (np *Processor) ListenerNotaryParsers() []event.NotaryParserInfo {
 	// new peer
 	p.SetRequestType(netmapEvent.AddPeerNotaryEvent)
 	p.SetParser(netmapEvent.ParseAddPeerNotary)
+	pp = append(pp, p)
+
+	// new node
+	p.SetRequestType(netmapEvent.AddNodeNotaryEvent)
+	p.SetParser(netmapEvent.ParseAddNodeNotary)
 	pp = append(pp, p)
 
 	// update state
@@ -207,7 +217,7 @@ func (np *Processor) ListenerNotaryHandlers() []event.NotaryHandlerInfo {
 	var (
 		h event.NotaryHandlerInfo
 
-		hh = make([]event.NotaryHandlerInfo, 0, 2)
+		hh = make([]event.NotaryHandlerInfo, 0, 3)
 	)
 
 	h.SetScriptHash(np.netmapClient.ContractAddress())
@@ -215,6 +225,11 @@ func (np *Processor) ListenerNotaryHandlers() []event.NotaryHandlerInfo {
 	// new peer
 	h.SetRequestType(netmapEvent.AddPeerNotaryEvent)
 	h.SetHandler(np.handleAddPeer)
+	hh = append(hh, h)
+
+	// new node
+	h.SetRequestType(netmapEvent.AddNodeNotaryEvent)
+	h.SetHandler(np.handleAddNode)
 	hh = append(hh, h)
 
 	// update state