+++ pkg/morph/client/netmap/netmap.go
@@ -3,27 +3,52 @@ package netmap
 import (
 	"errors"
 	"fmt"
+	"math/big"
 
+	"github.com/google/uuid"
+	"github.com/nspcc-dev/neo-go/pkg/neorpc/result"
+	"github.com/nspcc-dev/neo-go/pkg/rpcclient/invoker"
 	"github.com/nspcc-dev/neo-go/pkg/vm/stackitem"
 	netmaprpc "github.com/nspcc-dev/neofs-contract/rpc/netmap"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client"
+	netmapEvent "github.com/nspcc-dev/neofs-node/pkg/morph/event/netmap"
 	"github.com/nspcc-dev/neofs-sdk-go/netmap"
 )
 
 // GetNetMapByEpoch calls "snapshotByEpoch" method with the given epoch and
 // decodes netmap.NetMap from the response.
 func (c *Client) GetNetMapByEpoch(epoch uint64) (*netmap.NetMap, error) {
-	invokePrm := client.TestInvokePrm{}
-	invokePrm.SetMethod(epochSnapshotMethod)
-	invokePrm.SetArgs(epoch)
+	var (
+		err error
+		nm  *netmap.NetMap
+	)
+	if !c.nodeV2 {
+		var res []stackitem.Item
 
-	res, err := c.client.TestInvoke(invokePrm)
-	if err != nil {
-		return nil, fmt.Errorf("could not perform test invocation (%s): %w",
-			epochSnapshotMethod, err)
+		invokePrm := client.TestInvokePrm{}
+		invokePrm.SetMethod(epochSnapshotMethod)
+		invokePrm.SetArgs(epoch)
+
+		res, err = c.client.TestInvoke(invokePrm)
+		if err != nil {
+			return nil, fmt.Errorf("could not perform test invocation (%s): %w",
+				epochSnapshotMethod, err)
+		}
+
+		nm, err = DecodeNetMap(res)
+	} else {
+		var (
+			inv    = invoker.New(c.client.Morph(), nil)
+			iter   result.Iterator
+			reader = netmaprpc.NewReader(inv, c.contract)
+			sess   uuid.UUID
+		)
+		sess, iter, err = reader.ListNodes2(big.NewInt(int64(epoch)))
+		if err != nil {
+			return nil, err
+		}
+		nm, err = collectNetmap(inv, sess, &iter)
 	}
-
-	nm, err := DecodeNetMap(res)
 	if err != nil {
 		return nil, err
 	}
@@ -36,33 +61,104 @@ func (c *Client) GetNetMapByEpoch(epoch uint64) (*netmap.NetMap, error) {
 // GetCandidates calls "netmapCandidates" method and decodes []netmap.NodeInfo
 // from the response.
 func (c *Client) GetCandidates() ([]netmap.NodeInfo, error) {
-	invokePrm := client.TestInvokePrm{}
-	invokePrm.SetMethod(netMapCandidatesMethod)
+	if !c.nodeV2 {
+		invokePrm := client.TestInvokePrm{}
+		invokePrm.SetMethod(netMapCandidatesMethod)
 
-	res, err := c.client.TestInvoke(invokePrm)
+		res, err := c.client.TestInvoke(invokePrm)
+		if err != nil {
+			return nil, fmt.Errorf("could not perform test invocation (%s): %w", netMapCandidatesMethod, err)
+		}
+
+		if len(res) > 0 {
+			return decodeNodeList(res[0])
+		}
+
+		return nil, nil
+	}
+	var (
+		inv             = invoker.New(c.client.Morph(), nil)
+		reader          = netmaprpc.NewReader(inv, c.contract)
+		sess, iter, err = reader.ListCandidates()
+	)
 	if err != nil {
-		return nil, fmt.Errorf("could not perform test invocation (%s): %w", netMapCandidatesMethod, err)
+		return nil, err
 	}
-
-	if len(res) > 0 {
-		return decodeNodeList(res[0])
-	}
-
-	return nil, nil
+	return CollectNodes(inv, sess, &iter, netmapEvent.Candidate2Info)
 }
 
-// NetMap calls "netmap" method and decode netmap.NetMap from the response.
+// NetMap calls "netmap" method (or listNodes for v2 nodes) and decodes
+// netmap.NetMap from the response.
 func (c *Client) NetMap() (*netmap.NetMap, error) {
-	invokePrm := client.TestInvokePrm{}
-	invokePrm.SetMethod(netMapMethod)
+	if !c.nodeV2 {
+		invokePrm := client.TestInvokePrm{}
+		invokePrm.SetMethod(netMapMethod)
 
-	res, err := c.client.TestInvoke(invokePrm)
-	if err != nil {
-		return nil, fmt.Errorf("could not perform test invocation (%s): %w",
-			netMapMethod, err)
+		res, err := c.client.TestInvoke(invokePrm)
+		if err != nil {
+			return nil, fmt.Errorf("could not perform test invocation (%s): %w",
+				netMapMethod, err)
+		}
+
+		return DecodeNetMap(res)
 	}
+	var (
+		inv             = invoker.New(c.client.Morph(), nil)
+		reader          = netmaprpc.NewReader(inv, c.contract)
+		sess, iter, err = reader.ListNodes()
+	)
+	if err != nil {
+		return nil, err
+	}
+	return collectNetmap(inv, sess, &iter)
+}
 
-	return DecodeNetMap(res)
+func collectNetmap(inv *invoker.Invoker, sess uuid.UUID, iter *result.Iterator) (*netmap.NetMap, error) {
+	nodes, err := CollectNodes(inv, sess, iter, netmapEvent.Node2Info)
+	if err != nil {
+		return nil, err
+	}
+	var nm = new(netmap.NetMap)
+	if len(nodes) > 0 {
+		nm.SetNodes(nodes)
+	}
+	return nm, nil
+}
+
+// CollectNodes gathers all node data from the provided iterator and closes it.
+func CollectNodes[N any, P interface {
+	*N
+	stackitem.Convertible
+}](inv *invoker.Invoker, sess uuid.UUID, iter *result.Iterator,
+	converter func(P) (netmap.NodeInfo, error)) ([]netmap.NodeInfo, error) {
+	var nodes []netmap.NodeInfo
+
+	defer func() {
+		_ = inv.TerminateSession(sess)
+	}()
+	items, err := inv.TraverseIterator(sess, iter, 0)
+	for err == nil && len(items) > 0 {
+		for _, itm := range items {
+			var (
+				n2  N
+				ni  netmap.NodeInfo
+				err = P(&n2).FromStackItem(itm)
+			)
+			if err != nil {
+				return nil, err
+			}
+			ni, err = converter(P(&n2))
+			if err != nil {
+				return nil, err
+			}
+			nodes = append(nodes, ni)
+		}
+		items, err = inv.TraverseIterator(sess, iter, 0)
+	}
+	if err != nil {
+		return nil, err
+	}
+	return nodes, nil
 }
 
 func DecodeNetMap(resStack []stackitem.Item) (*netmap.NetMap, error) {